+++ crates/protos/src/protos/mod.rs
@@ -1277,10 +1277,10 @@ pub mod coresdk {
                         workflow_activation_job::Variant::ResolveNexusOperation(_) => {
                             write!(f, "ResolveNexusOperation")
                         }
-                        workflow_activation_job::Variant::DeliverStreamMessages(d) => {
+                        workflow_activation_job::Variant::DeliverStreamRecords(d) => {
                             write!(
                                 f,
-                                "DeliverStreamMessages({}, {}..{})",
+                                "DeliverStreamRecords({}, {}..{})",
                                 d.stream_id, d.from_offset, d.to_offset
                             )
                         }
@@ -1499,13 +1499,13 @@ pub mod coresdk {
                 }
             }
 
-            impl Display for AddStreamMessages {
+            impl Display for AppendStreamRecords {
                 fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
                     write!(
                         f,
-                        "AddStreamMessages({}, {} messages)",
+                        "AppendStreamRecords({}, {} records)",
                         self.stream_id,
-                        self.messages.len()
+                        self.records.len()
                     )
                 }
             }
@@ -1852,8 +1852,8 @@ pub mod temporal {
                             Attributes::SubscribeStreamCommandAttributes(_) => {
                                 CommandType::SubscribeStream
                             }
-                            Attributes::AddStreamMessagesCommandAttributes(_) => {
-                                CommandType::AddStreamMessages
+                            Attributes::AppendStreamRecordsCommandAttributes(_) => {
+                                CommandType::AppendStreamRecords
                             }
                             Attributes::CompleteWorkflowExecutionCommandAttributes(_) => {
                                 CommandType::CompleteWorkflowExecution
@@ -1908,12 +1908,12 @@ pub mod temporal {
                         }
                     }
 
-                    impl From<workflow_commands::AddStreamMessages> for command::Attributes {
-                        fn from(s: workflow_commands::AddStreamMessages) -> Self {
-                            Self::AddStreamMessagesCommandAttributes(
-                                AddStreamMessagesCommandAttributes {
+                    impl From<workflow_commands::AppendStreamRecords> for command::Attributes {
+                        fn from(s: workflow_commands::AppendStreamRecords) -> Self {
+                            Self::AppendStreamRecordsCommandAttributes(
+                                AppendStreamRecordsCommandAttributes {
                                     stream_id: s.stream_id,
-                                    messages: s.messages,
+                                    records: s.records,
                                 },
                             )
                         }
@@ -2390,7 +2390,7 @@ pub mod temporal {
                                 | EventType::UpsertWorkflowSearchAttributes
                                 | EventType::WorkflowPropertiesModified
                                 | EventType::WorkflowStreamSubscribed
-                                | EventType::WorkflowStreamMessagesAdded
+                                | EventType::WorkflowStreamRecordsAppended
                                 | EventType::NexusOperationScheduled
                                 | EventType::NexusOperationCancelRequested
                                 | EventType::WorkflowExecutionCanceled
@@ -2491,7 +2491,7 @@ pub mod temporal {
                             if let Some(a) = self.attributes.as_ref() {
                                 match a {
                                     Attributes::WorkflowStreamSubscribedEventAttributes(_) => false,
-                                    Attributes::WorkflowStreamMessagesAddedEventAttributes(_) => {
+                                    Attributes::WorkflowStreamRecordsAppendedEventAttributes(_) => {
                                         false
                                     }
                                     Attributes::WorkflowExecutionStartedEventAttributes(_) => false,
@@ -2582,7 +2582,7 @@ pub mod temporal {
                             // I just absolutely _love_ this
                             match self {
                             Attributes::WorkflowStreamSubscribedEventAttributes(_) => { EventType::WorkflowStreamSubscribed }
-                            Attributes::WorkflowStreamMessagesAddedEventAttributes(_) => { EventType::WorkflowStreamMessagesAdded }
+                            Attributes::WorkflowStreamRecordsAppendedEventAttributes(_) => { EventType::WorkflowStreamRecordsAppended }
                             Attributes::WorkflowExecutionStartedEventAttributes(_) => { EventType::WorkflowExecutionStarted }
                             Attributes::WorkflowExecutionCompletedEventAttributes(_) => { EventType::WorkflowExecutionCompleted }
                             Attributes::WorkflowExecutionFailedEventAttributes(_) => { EventType::WorkflowExecutionFailed }