+++ crates/sdk-core/src/worker/workflow/mod.rs
@@ -1530,7 +1530,7 @@ enum WFCommandVariant {
     ScheduleNexusOperation(ScheduleNexusOperation),
     RequestCancelNexusOperation(RequestCancelNexusOperation),
     SubscribeStream(SubscribeStream),
-    AddStreamMessages(AddStreamMessages),
+    AppendStreamRecords(AppendStreamRecords),
 }
 
 impl TryFrom<WorkflowCommand> for WFCommand {
@@ -1540,8 +1540,8 @@ impl TryFrom<WorkflowCommand> for WFCommand {
         let variant = match c.variant.ok_or(EmptyWorkflowCommandErr)? {
             workflow_command::Variant::StartTimer(s) => WFCommandVariant::AddTimer(s),
             workflow_command::Variant::SubscribeStream(s) => WFCommandVariant::SubscribeStream(s),
-            workflow_command::Variant::AddStreamMessages(s) => {
-                WFCommandVariant::AddStreamMessages(s)
+            workflow_command::Variant::AppendStreamRecords(s) => {
+                WFCommandVariant::AppendStreamRecords(s)
             }
             workflow_command::Variant::CancelTimer(s) => WFCommandVariant::CancelTimer(s),
             workflow_command::Variant::ScheduleActivity(s) => WFCommandVariant::AddActivity(s),